linker-rc2

+++ src/app/main/components/category-tree/category-tree.component.ts
@@ -4,10 +4,11 @@ import { LgMenuItem } from '../../../shared/modules/lg-context-menu/interfaces/l
 import { LinkService } from '../../../main/services/link.service';
 import { Category } from '../../models/category';
 import { AddCategoryDialog } from '../../dialogs/categories/add-category/add-category.component';
-import { MatDialog } from '@angular/material';
+import { MatDialog, MatSnackBar } from '@angular/material';
 import { EditCategoryDialog } from '../../dialogs/categories/edit-category/edit-category.component';
 import { SendShareRequestDialog } from '../../dialogs/share/send-share-request/send-share-request.component';
-import { MoveCategoryDialog } from '../../dialogs/categories/move-category/move-category.component';
+import { LinkItem } from '../../models/link-item';
+import { CategoryCacheService } from '../../services/category-cache.service';
 
 @Component({
   selector: 'category-tree',
@@ -15,7 +16,7 @@ import { MoveCategoryDialog } from '../../dialogs/categories/move-category/move-
   styleUrls: ['./category-tree.component.scss']
 })
 export class CategoryTreeComponent implements OnInit {
-
+  rootExpanded: boolean = true;
   rootCategories: Category[] = [];
   menuItems: LgMenuItem[] = [
     {
@@ -24,7 +25,7 @@ export class CategoryTreeComponent implements OnInit {
       action: (category: Category) => {
         this.dialog.open(AddCategoryDialog, {
           height: '300px',
-          width: '350px',
+          width: '450px',
           data: {
             'parentCategoryId': category.id
           }
@@ -37,18 +38,7 @@ export class CategoryTreeComponent implements OnInit {
       action: (category: Category) => {
         this.dialog.open(SendShareRequestDialog, {
           height: '300px',
-          width: '350px',
-          data: category
-        });
-      }
-    },
-    {
-      name: 'Move',
-      icon: 'swap_vert',
-      action: (category: Category) => {
-        this.dialog.open(MoveCategoryDialog, {
-          height: '300px',
-          width: '350px',
+          width: '450px',
           data: category
         });
       }
@@ -59,7 +49,7 @@ export class CategoryTreeComponent implements OnInit {
       action: (category: Category) => {
         this.dialog.open(EditCategoryDialog, {
           height: '300px',
-          width: '350px',
+          width: '450px',
           data: category
         });
       }
@@ -68,7 +58,9 @@ export class CategoryTreeComponent implements OnInit {
       name: 'Delete',
       icon: 'delete',
       action: (category: Category) => {
-        this.categoryService.deleteLinkCategory(category.id);
+        if (category !== null) {
+          this.categoryService.deleteLinkCategory(category.id);
+        }
       }
     }
     
@@ -77,7 +69,9 @@ export class CategoryTreeComponent implements OnInit {
   constructor(
     private categoryService: CategoryService,
     private linkService: LinkService,
-    private dialog: MatDialog
+    private dialog: MatDialog,
+    private categoryCacheService: CategoryCacheService,
+    private snackBar: MatSnackBar,
   ) { }
 
   ngOnInit() {
@@ -86,7 +80,11 @@ export class CategoryTreeComponent implements OnInit {
 
   getCategories() {
     this.categoryService.getRootCategories$().subscribe((categories: Category[]) => {
-      this.rootCategories = categories;
+      this.rootCategories = categories.sort(function(a, b){
+        if(a.title.toLowerCase() < b.title.toLowerCase()) return -1;
+        if(a.title.toLowerCase() > b.title.toLowerCase()) return 1;
+        return 0;
+      })
     });
   }
 
@@ -94,6 +92,41 @@ export class CategoryTreeComponent implements OnInit {
     category.expanded = !category.expanded;
     this.linkService.friendId.next(null);
     this.linkService.category.next(category);
+  }
+
+  onDrop(event: any, receivingCatgory: Category = null) {
+    let droppedItem: any = event.dragData;
+
+    // Link was dropped
+    if (droppedItem.linkUrl !== undefined) {
+      if (droppedItem.categoryId !== receivingCatgory.id) {
+        this.linkService.moveLink(droppedItem, receivingCatgory.id);
+      }
+    }
+    // Category was dropped
+    else if (droppedItem.title !== undefined) {
+      if (receivingCatgory === null && droppedItem.parentCategoryId) {
+        this.categoryService.moveCategory(droppedItem.id, null);
+      }
+      else {
+        if (droppedItem.id !== receivingCatgory.id && droppedItem.parentCategoryId !== receivingCatgory.id) {
+          if (!this.categoryCacheService.isParentOfCategory(droppedItem.id, receivingCatgory.id)) {
+            this.categoryService.moveCategory(droppedItem.id, receivingCatgory.id);
+          }
+          else {
+            this.snackBar.open("Cannot move category into its' child", '', {
+              duration: 2000,
+            });
+          }
+        }
+      }
+    } 
+  }
+
+  toggleRoot() {
+    this.rootExpanded = !this.rootExpanded;
+    this.linkService.friendId.next(null);
+    this.linkService.category.next(null);
   }
 
 }
